# contracts/openapi/lifetrace-sync-v1.json
{
  "components": {
    "schemas": {
...
      "ApiErrorV1": {
        "description": "Uniform API error body returned by sync endpoints.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JsonValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "fieldErrors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RequestId"
              },
              {
                "type": "null"
              }
            ]
          },
          "retryable": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "type": "object"
      },
...
      "AppId": {
        "description": "Client application id. Known values: lifetrace-desktop, lifetrace-web, lifetrace-*-android.",
        "type": "string"
      },
...
      "CapabilitiesResponseV1": {
        "description": "`GET /api/v1/sync/capabilities` response.",
        "properties": {
          "maximumAtomicGroupSize": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "maximumPullBatchSize": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "maximumPushBatchSize": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "maximumRequestBytes": {
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "maximumSnapshotPageSize": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "minimumClientVersions": {
            "items": {
              "$ref": "#/components/schemas/MinimumClientVersion"
            },
            "type": "array"
          },
          "minimumSchemaVersion": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "protocolVersion": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "schemaVersion": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "serverTime": {
            "format": "date-time",
            "type": "string"
          },
          "supportedEntityTypes": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "supportedProtocolVersions": {
            "items": {
              "format": "uint32",
              "minimum": 0,
              "type": "integer"
            },
            "type": "array"
          },
          "tombstoneRetentionDays": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "protocolVersion",
          "supportedProtocolVersions",
          "schemaVersion",
          "minimumSchemaVersion",
          "minimumClientVersions",
          "maximumPushBatchSize",
          "maximumPullBatchSize",
          "maximumRequestBytes",
          "maximumSnapshotPageSize",
          "maximumAtomicGroupSize",
          "tombstoneRetentionDays",
          "supportedEntityTypes",
          "serverTime"
        ],
        "type": "object"
      },
...
      "ChangeOperation": {
        "description": "Change operation. Known values: upsert, delete. Unknown values are preserved.",
        "type": "string"
      },
...
      "ClientPlatform": {
        "description": "Client platform. Known values: windows, android, web. Unknown values are preserved.",
        "type": "string"
      },
...
      "Cursor": {
        "description": "Opaque server cursor. Clients must never parse, add, subtract or guess it.\nWire format: string.",
        "type": "string"
      },
...
      "DeviceId": {
        "description": "Stable device identifier (string, usually UUID v4).",
        "type": "string"
      },
...
      "EntityId": {
        "description": "Entity identifier. New IDs are UUID v4; legacy non-UUID IDs are preserved verbatim.",
        "type": "string"
      },
...
      "EntitySnapshotV1": {
        "description": "One current entity in a snapshot.",
        "properties": {
          "entityId": {
            "$ref": "#/components/schemas/EntityId"
          },
          "entityType": {
            "$ref": "#/components/schemas/EntityType"
          },
          "payload": {
            "$ref": "#/components/schemas/JsonValue"
          },
          "serverVersion": {
            "$ref": "#/components/schemas/ServerVersion"
          }
        },
        "required": [
          "entityType",
          "entityId",
          "serverVersion",
          "payload"
        ],
        "type": "object"
      },
      "EntityType": {
        "description": "Registered entity type. Known values: identity.user, identity.device, finance.account, finance.category, finance.transaction, finance.transaction_evidence, habit.activity, habit.log, review.daily, note.folder, note.note, note.tag, note.tag_relation, note.relation, note.revision, english.article, english.learning_record, english.highlight, english.note, english.vocabulary, english.vocabulary_occurrence, english.vocabulary_review_state, workout.import, workout.workout, workout.exercise, workout.set, workout.training_note, file.metadata, entity.link, user.preference. Unknown values are preserved.",
        "type": "string"
      },
      "ErrorCode": {
        "description": "Stable error code. Unknown codes from newer servers are preserved verbatim.",
        "enum": [
          "LIFETRACE_PROTOCOL_UNSUPPORTED",
          "LIFETRACE_SCHEMA_UNSUPPORTED",
          "LIFETRACE_CLIENT_TOO_OLD",
          "LIFETRACE_APP_ID_UNSUPPORTED",
          "LIFETRACE_AUTH_REQUIRED",
          "LIFETRACE_AUTH_INVALID",
          "LIFETRACE_AUTH_PASSWORD_INVALID",
          "LIFETRACE_AUTH_USER_DISABLED",
          "LIFETRACE_AUTH_USER_LOCKED",
          "LIFETRACE_AUTH_ACCESS_TOKEN_EXPIRED",
          "LIFETRACE_AUTH_REFRESH_TOKEN_EXPIRED",
          "LIFETRACE_AUTH_REFRESH_TOKEN_REUSED",
          "LIFETRACE_AUTH_SESSION_REVOKED",
          "LIFETRACE_AUTH_DEVICE_REVOKED",
          "LIFETRACE_AUTH_APP_REVOKED",
          "LIFETRACE_AUTH_SCOPE_DENIED",
          "LIFETRACE_AUTH_PASSWORD_POLICY_FAILED",
          "LIFETRACE_AUTH_PASSWORD_RESET_INVALID",
          "LIFETRACE_AUTH_PASSWORD_RESET_EXPIRED",
          "LIFETRACE_AUTH_RATE_LIMITED",
          "LIFETRACE_AUTH_CSRF_INVALID",
          "LIFETRACE_AUTH_REGISTRATION_DISABLED",
          "LIFETRACE_AUTH_INVITE_INVALID",
          "LIFETRACE_DEVICE_NOT_REGISTERED",
          "LIFETRACE_DEVICE_REVOKED",
          "LIFETRACE_INVALID_REQUEST",
          "LIFETRACE_BATCH_TOO_LARGE",
          "LIFETRACE_PAYLOAD_TOO_LARGE",
          "LIFETRACE_UNKNOWN_ENTITY_TYPE",
          "LIFETRACE_INVALID_ENTITY_PAYLOAD",
          "LIFETRACE_DEPENDENCY_MISSING",
          "LIFETRACE_CHANGE_ID_REUSE",
          "LIFETRACE_BASE_VERSION_MISMATCH",
          "LIFETRACE_CURSOR_INVALID",
          "LIFETRACE_CURSOR_EXPIRED",
          "LIFETRACE_SNAPSHOT_REQUIRED",
          "LIFETRACE_ATOMIC_GROUP_FAILED",
          "LIFETRACE_RATE_LIMITED",
          "LIFETRACE_TEMPORARILY_UNAVAILABLE",
          "LIFETRACE_INTERNAL_ERROR"
        ],
        "type": "string"
      },
      "FieldError": {
        "description": "Field-level validation error inside `ApiErrorV1`.",
        "properties": {
          "code": {
            "description": "Field-level error code (string; server-defined, forward compatible).",
            "type": "string"
          },
          "field": {
            "description": "Field path, for example `changes[3].entityType`.",
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "JsonValue": true,
...
      "MinimumClientVersion": {
        "description": "Minimum required client version for one app id.",
        "properties": {
          "appId": {
            "$ref": "#/components/schemas/AppId"
          },
          "clientVersion": {
            "type": "string"
          }
        },
        "required": [
          "appId",
          "clientVersion"
        ],
        "type": "object"
      },
...
      "PullRequestV1": {
        "description": "`POST /api/v1/sync/pull` request.",
        "properties": {
          "afterCursor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Cursor"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque server cursor. `None` means \"start from the beginning\" (may be\nanswered with `LIFETRACE_SNAPSHOT_REQUIRED` when history is pruned)."
          },
          "client": {
            "$ref": "#/components/schemas/SyncClientInfo"
          },
          "entityTypes": {
            "description": "Optional entity type filter.",
            "items": {
              "$ref": "#/components/schemas/EntityType"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "limit": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "requestId": {
            "$ref": "#/components/schemas/RequestId"
          }
        },
        "required": [
          "requestId",
          "client",
          "limit"
        ],
        "type": "object"
      },
      "PullResponseV1": {
        "description": "`POST /api/v1/sync/pull` response.\n\nChanges are ordered strictly by server cursor. Clients MUST apply them in\norder and MUST NOT re-sort by `updatedAt`; only persist `next_cursor` after\nthe whole batch succeeded.",
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/ServerChangeV1"
            },
            "type": "array"
          },
          "hasMore": {
            "type": "boolean"
          },
          "nextCursor": {
            "$ref": "#/components/schemas/Cursor"
          },
          "requestId": {
            "$ref": "#/components/schemas/RequestId"
          },
          "serverTime": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "serverTime",
          "changes",
          "nextCursor",
          "hasMore"
        ],
        "type": "object"
      },
...
      "PushRequestV1": {
        "description": "`POST /api/v1/sync/push` request.",
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/SyncChangeV1"
            },
            "type": "array"
          },
          "client": {
            "$ref": "#/components/schemas/SyncClientInfo"
          },
          "requestId": {
            "$ref": "#/components/schemas/RequestId",
            "description": "Tracing only; never replaces `changeId` for idempotency."
          }
        },
        "required": [
          "requestId",
          "client",
          "changes"
        ],
        "type": "object"
      },
      "PushResponseV1": {
        "description": "`POST /api/v1/sync/push` response.",
        "properties": {
          "latestCursor": {
            "$ref": "#/components/schemas/Cursor"
          },
          "requestId": {
            "$ref": "#/components/schemas/RequestId"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/PushChangeResultV1"
            },
            "type": "array"
          },
          "serverTime": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "serverTime",
          "results",
          "latestCursor"
        ],
        "type": "object"
      },
...
      "RequestId": {
        "description": "Request tracing id. It never replaces changeId for idempotency.",
        "type": "string"
      },
...
      "ServerChangeV1": {
        "description": "A change as served by pull, in strict server cursor order.",
        "properties": {
          "cursor": {
            "$ref": "#/components/schemas/Cursor"
          },
          "entityId": {
            "$ref": "#/components/schemas/EntityId"
          },
          "entityType": {
            "$ref": "#/components/schemas/EntityType"
          },
          "operation": {
            "$ref": "#/components/schemas/ChangeOperation"
          },
          "originDeviceId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceId"
              },
              {
                "type": "null"
              }
            ]
          },
          "payload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JsonValue"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present for `upsert`; empty for `delete`."
          },
          "serverModifiedAt": {
            "format": "date-time",
            "type": "string"
          },
          "serverVersion": {
            "$ref": "#/components/schemas/ServerVersion"
          },
          "tombstone": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TombstoneV1"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present for `delete` (and for conflicts involving deletion)."
          }
        },
        "required": [
          "cursor",
          "entityType",
          "entityId",
          "operation",
          "serverVersion",
          "serverModifiedAt"
        ],
        "type": "object"
      },
...
      "SnapshotId": {
        "description": "Stable id assigned by the server to a snapshot; all pages share one consistent view.",
        "type": "string"
      },
      "SnapshotRequestV1": {
        "description": "`POST /api/v1/sync/snapshot` request.",
        "properties": {
          "client": {
            "$ref": "#/components/schemas/SyncClientInfo"
          },
          "entityTypes": {
            "items": {
              "$ref": "#/components/schemas/EntityType"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "pageSize": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "pageToken": {
            "type": [
              "string",
              "null"
            ]
          },
          "requestId": {
            "$ref": "#/components/schemas/RequestId"
          },
          "snapshotId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SnapshotId"
              },
              {
                "type": "null"
              }
            ],
            "description": "Omitted on the first page; the server assigns one and expects it on\nsubsequent pages so all pages share one consistent view."
          }
        },
        "required": [
          "requestId",
          "client",
          "pageSize"
        ],
        "type": "object"
      },
      "SnapshotResponseV1": {
        "description": "`POST /api/v1/sync/snapshot` response.\n\nAll pages of one snapshot correspond to one consistent view. After the\nclient finishes (completed == true), it sets its cursor to\n`snapshot_cursor` and continues with Pull from there; concurrent changes\nare then received without gaps.",
        "properties": {
          "completed": {
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EntitySnapshotV1"
            },
            "type": "array"
          },
          "nextPageToken": {
            "type": [
              "string",
              "null"
            ]
          },
          "requestId": {
            "$ref": "#/components/schemas/RequestId"
          },
          "serverTime": {
            "format": "date-time",
            "type": "string"
          },
          "snapshotCursor": {
            "$ref": "#/components/schemas/Cursor"
          },
          "snapshotId": {
            "$ref": "#/components/schemas/SnapshotId"
          }
        },
        "required": [
          "requestId",
          "snapshotId",
          "snapshotCursor",
          "items",
          "completed",
          "serverTime"
        ],
        "type": "object"
      },
...
      "SyncClientInfo": {
        "description": "Sync client info carried by every request body (headers may also carry it;\nthe server validates consistency).",
        "properties": {
          "appId": {
            "$ref": "#/components/schemas/AppId"
          },
          "clientVersion": {
            "description": "Client app release version (for example `0.2.1`).",
            "type": "string"
          },
          "deviceId": {
            "$ref": "#/components/schemas/DeviceId"
          },
          "platform": {
            "$ref": "#/components/schemas/ClientPlatform"
          },
          "protocolVersion": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          },
          "schemaVersion": {
            "format": "uint32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "appId",
          "clientVersion",
          "platform",
          "protocolVersion",
          "schemaVersion",
          "deviceId"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "GENERATED FILE - DO NOT EDIT MANUALLY.\nPublic sync contract v1 for LifeTrace. Rust types in crates/lifetrace-contracts are the authoritative source.\n\nVersioning: protocolVersion 1, schemaVersion 1. Unknown JSON fields are ignored; unknown enum values are preserved as strings. Cursor/serverVersion are opaque strings on the wire. Money is integer cents (amountCents). Timestamps are RFC3339 UTC; natural days are YYYY-MM-DD.\n\nPush idempotency key: userId + changeId. Same changeId + same payload replays the first result; same changeId + different payload returns LIFETRACE_CHANGE_ID_REUSE.\n\nConflicts are explicit (no automatic last-write-wins): when baseServerVersion != current serverVersion the server returns conflict with the current entity or tombstone. Resolution is client-side: keep_server, keep_local (new changeId), manual_merge.\n\nDeletes produce tombstones. Re-creating a deleted entity requires a NEW entity id; restoring a soft delete must be based on the tombstone's latest serverVersion.\n\nSnapshot pages share one consistent view; after completion set the client cursor to snapshotCursor and continue with Pull.",
    "title": "LifeTrace Sync Protocol v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/sync/capabilities": {
      "get": {
        "description": "Clients should call this first to validate versions, batch limits and supported entity types.",
        "operationId": "syncCapabilities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponseV1"
                }
              }
            },
            "description": "Capabilities"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "426": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          }
        },
        "summary": "Return protocol and server capabilities"
      }
    },
    "/api/v1/sync/pull": {
      "post": {
        "description": "Responses are ordered strictly by server cursor. Clients must apply them in order and never re-sort by updatedAt; only persist nextCursor after the whole batch succeeded. An expired cursor returns LIFETRACE_CURSOR_EXPIRED and the client must run a snapshot first.",
        "operationId": "syncPull",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PullRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PullResponseV1"
                }
              }
            },
            "description": "Ordered changes"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "426": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          }
        },
        "summary": "Pull changes after a cursor"
      }
    },
    "/api/v1/sync/push": {
      "post": {
        "description": "Each change returns accepted, duplicate, conflict or rejected. Business conflicts do not fail the whole request. Changes sharing an atomicGroupId must be in this request and succeed or fail together. Batch limits are declared by capabilities.",
        "operationId": "syncPush",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushResponseV1"
                }
              }
            },
            "description": "Per-change results"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "426": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          }
        },
        "summary": "Push a batch of changes"
      }
    },
    "/api/v1/sync/snapshot": {
      "post": {
        "description": "All pages of one snapshot correspond to one consistent view. After the final page set the client cursor to snapshotCursor, then continue with Pull to receive concurrent changes without gaps.",
        "operationId": "syncSnapshot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponseV1"
                }
              }
            },
            "description": "Snapshot page"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "426": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorV1"
                }
              }
            },
            "description": "Uniform API error body with a stable ErrorCode"
          }
        },
        "summary": "Fetch a consistent snapshot of current entities"
      }
    }
  }
}